Bundler is not required to run the aws-rotate-keys anymore...
... also extract CLI to follow gem conventions
