+++ exe/aws-rotate-keys
@@ -1,6 +1,5 @@
 #!/usr/bin/env ruby
 
-require "bundler/setup"
 require "aws_rotate_keys"
 
-AwsRotateKeys.call
+AwsRotateKeys::CLI.call
