+++ lib/aws_rotate_keys/version.rb
